+++ src/datannurpy/scanner/statistical.py
@@ -11,7 +11,7 @@ import numpy as np
 import pandas as pd
 import pyarrow as pa
 
-from ..entities import Variable
+from ..schema import Variable
 from .utils import build_variables